+++ tests/unit/with_db/test_scopes.py
@@ -67,8 +67,9 @@ class TestAlert:
 
         alert = Alert("delete_me")
 
-        assert alert.AlertId == "delete_me"
         assert alert.Action == "Random action"
         assert alert.Info == "Random info"
-        assert alert.AlertLabel == "delete_me"
+        assert alert.Label == "delete_me"
+        assert alert.Title == "Delete Me"
+        assert alert.User == "movai@internal"
         assert hasattr(alert, "LastUpdate")