+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# vTBD
+- [BP-1519](https://movai.atlassian.net/browse/BP-1519): Edit Alert schema
+
 # v3.10.0
 - [BP-1521](https://movai.atlassian.net/browse/BP-1521): Logic to activate and deactivate alert
 - [BP-1520](https://movai.atlassian.net/browse/BP-1520): Deactivate alerts on start / stop flow